Made first level with 2 mages and changed the attack for the player
the player now just uses a collider out in front of them to detect and do damage to enemies. removed the player model and just kept the sword. no animation for the player attack anymore
