+++ Prototype/Assets/Map Assets/Prefabs/HealthPickup.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 1cfde5d7e2ee9fe4f94d02c91c758b03
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
