+++ Prototype/Assets/Map Assets/Prefabs/Mage.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 20705ea9e6835d842ab7dcddaa4f3948
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
