+++ Prototype/Assets/Map Assets/Prefabs/Map Prefabs.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 20e99dd65baf46f4b8e778b5e56456bb
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
