+++ Prototype/Assets/Map Assets/Prefabs/Map Prefabs/bookcaseWideFilled.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 4b45c0d895992804194db13815c55dfb
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
