+++ Prototype/Assets/Map Assets/Prefabs/Map Prefabs/tableLarge.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 5e77af32211084a4d8d254c5e3989016
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
