+++ Prototype/Assets/Map Assets/Prefabs/Map Prefabs/wallDecorationA.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 04aa83ba31df22c43b141b4850906638
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
