+++ Prototype/Assets/Map Assets/Prefabs/Map Prefabs/wall_end_broken.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 87fff42c6bca3094eaddd88ba9faa558
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
