+++ Prototype/Assets/Map Assets/Prefabs/Map Prefabs/wall_window.prefab
@@ -1,0 +1,88 @@
+%YAML 1.1
+%TAG !u! tag:unity3d.com,2011:
+--- !u!1001 &151639094429009247
+PrefabInstance:
+  m_ObjectHideFlags: 0
+  serializedVersion: 2
+  m_Modification:
+    serializedVersion: 3
+    m_TransformParent: {fileID: 0}
+    m_Modifications:
+    - target: {fileID: -8679921383154817045, guid: 79404516b83ce7d49aae24e5560dbaf4, type: 3}
+      propertyPath: m_LocalPosition.x
+      value: 15.549
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 79404516b83ce7d49aae24e5560dbaf4, type: 3}
+      propertyPath: m_LocalPosition.y
+      value: -22.26371
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 79404516b83ce7d49aae24e5560dbaf4, type: 3}
+      propertyPath: m_LocalPosition.z
+      value: -8.141
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 79404516b83ce7d49aae24e5560dbaf4, type: 3}
+      propertyPath: m_LocalRotation.w
+      value: 1
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 79404516b83ce7d49aae24e5560dbaf4, type: 3}
+      propertyPath: m_LocalRotation.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 79404516b83ce7d49aae24e5560dbaf4, type: 3}
+      propertyPath: m_LocalRotation.y
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 79404516b83ce7d49aae24e5560dbaf4, type: 3}
+      propertyPath: m_LocalRotation.z
+      value: -0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 79404516b83ce7d49aae24e5560dbaf4, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.x
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 79404516b83ce7d49aae24e5560dbaf4, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.y
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: -8679921383154817045, guid: 79404516b83ce7d49aae24e5560dbaf4, type: 3}
+      propertyPath: m_LocalEulerAnglesHint.z
+      value: 0
+      objectReference: {fileID: 0}
+    - target: {fileID: 919132149155446097, guid: 79404516b83ce7d49aae24e5560dbaf4, type: 3}
+      propertyPath: m_Name
+      value: wall_window
+      objectReference: {fileID: 0}
+    m_RemovedComponents: []
+    m_RemovedGameObjects: []
+    m_AddedGameObjects: []
+    m_AddedComponents:
+    - targetCorrespondingSourceObject: {fileID: 919132149155446097, guid: 79404516b83ce7d49aae24e5560dbaf4, type: 3}
+      insertIndex: -1
+      addedObject: {fileID: 5334511079782820756}
+  m_SourcePrefab: {fileID: 100100000, guid: 79404516b83ce7d49aae24e5560dbaf4, type: 3}
+--- !u!1 &1070681479707490318 stripped
+GameObject:
+  m_CorrespondingSourceObject: {fileID: 919132149155446097, guid: 79404516b83ce7d49aae24e5560dbaf4, type: 3}
+  m_PrefabInstance: {fileID: 151639094429009247}
+  m_PrefabAsset: {fileID: 0}
+--- !u!65 &5334511079782820756
+BoxCollider:
+  m_ObjectHideFlags: 0
+  m_CorrespondingSourceObject: {fileID: 0}
+  m_PrefabInstance: {fileID: 0}
+  m_PrefabAsset: {fileID: 0}
+  m_GameObject: {fileID: 1070681479707490318}
+  m_Material: {fileID: 0}
+  m_IncludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_ExcludeLayers:
+    serializedVersion: 2
+    m_Bits: 0
+  m_LayerOverridePriority: 0
+  m_IsTrigger: 0
+  m_ProvidesContacts: 0
+  m_Enabled: 1
+  serializedVersion: 3
+  m_Size: {x: 6, y: 4, z: 1.5000007}
+  m_Center: {x: -0.00000011920929, y: 2, z: 0.0000002682209}
