+++ Prototype/Assets/Map Assets/Prefabs/Map Prefabs/wall_window.prefab.meta
@@ -1,0 +1,7 @@
+fileFormatVersion: 2
+guid: 38f58ebca8171544ebc9f9c266dc5da1
+PrefabImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
