+++ Prototype/Assets/Map Assets/Prefabs/Missle.prefab
@@ -59,6 +59,7 @@ MonoBehaviour:
   m_EditorClassIdentifier: 
   type: 2
   body: {fileID: 1687214239478801066}
+  playerAttack: {fileID: 0}
   damageAmount: 10
   damageRate: 0
   speed: 5
