--- Prototype/Assets/Scripts/manager/boosts.cs.meta
@@ -1,2 +1,0 @@
-fileFormatVersion: 2
-guid: 7a1b63a627593a54fba5d14459f941bd
