+++ Prototype/Assets/Scripts/manager/pickupItem.cs.meta
@@ -1,0 +1,2 @@
+fileFormatVersion: 2
+guid: fc22d6d80ef230c47bccf67bd67fe415
