+++ Prototype/Assets/Scripts/player/playerController.cs
@@ -19,6 +19,9 @@ public class playerController : MonoBehaviour, IDamage, IPickup
     public int HPOrig => HP;
     private int maxHP;
 
+    [SerializeField] int DEF;
+    private int maxDEF;
+
     [SerializeField] int speed;
     [SerializeField] int sprintMod;
